Moved DroidCameraPreview2 class to new file.
