+++ XFGetCameraData/XFGetCameraData.Android/XFGetCameraData.Android.csproj
@@ -68,6 +68,7 @@
     <Compile Include="CustomRenderers\CameraPreviewCallback.cs" />
     <Compile Include="CustomRenderers\CameraPreviewRenderer.cs" />
     <Compile Include="CustomRenderers\CameraPreviewRenderer2.cs" />
+    <Compile Include="CustomRenderers\DroidCameraPreview2.cs" />
     <Compile Include="CustomRenderers\Listeners\CameraCaptureListener.cs" />
     <Compile Include="CustomRenderers\Listeners\CameraCaptureStateListener.cs" />
     <Compile Include="CustomRenderers\Listeners\CameraCaptureStillPictureSessionListener.cs" />
